Because it will have dynamic questions of all kinds, we need to dynamically create them.
Instead of doing this with JavaScript I made a helper element called question-inflator that bundles away the logic for dynamically requesting an element, setting it up, and adding it to the page. Using this element allows me to declaratively use dom-repeat to add them all to the page, hiding away the complexity of actually doing that.